# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: EduEarn Student Productivity & Earning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/join-stream:
    post:
      operationId: joinStream
      tags: [users]
      summary: Join a class stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinStreamRequest"
      responses:
        "200":
          description: Joined stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /internships:
    get:
      operationId: getInternships
      tags: [internships]
      summary: Get internship/job listings
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [internship, part_time, freelance]
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of internships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InternshipListing"
    post:
      operationId: createInternshipListing
      tags: [internships]
      summary: Create internship listing (employer role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInternshipListingRequest"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternshipListing"
        "403":
          description: Employer role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /internships/{id}:
    delete:
      operationId: deleteInternshipListing
      tags: [internships]
      summary: Delete own internship listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /streams/{stream}/members:
    get:
      operationId: getStreamMembers
      tags: [streams]
      summary: Get members of a specific stream
      parameters:
        - name: stream
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stream members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StreamMember"
  /study/sessions:
    post:
      operationId: logStudySession
      tags: [study]
      summary: Log a completed study session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogStudySessionRequest"
      responses:
        "201":
          description: Session logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySessionResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, teacher, employer]
          default: student
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher, employer]
        points:
          type: integer
        streak:
          type: integer
        stream:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        lastCheckin:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, points, streak, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user]
...
    JoinStreamRequest:
      type: object
      properties:
        stream:
          type: string
      required: [stream]
...
    InternshipListing:
      type: object
      properties:
        id:
          type: integer
        employerId:
          type: integer
        employerName:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [internship, part_time, freelance]
        payment:
          type: integer
        paymentLabel:
          type: string
        location:
          type: string
        applyLink:
          type: string
          nullable: true
        applyEmail:
          type: string
          nullable: true
        skills:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, employerId, employerName, title, description, type, payment, paymentLabel, location, createdAt]
    CreateInternshipListingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [internship, part_time, freelance]
        payment:
          type: integer
        paymentLabel:
          type: string
        location:
          type: string
        applyLink:
          type: string
        applyEmail:
          type: string
        skills:
          type: string
      required: [title, description, type, payment, paymentLabel, location]
...
    StreamMember:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        points:
          type: integer
        streak:
          type: integer
        tasksCompleted:
          type: integer
        isCurrentUser:
          type: boolean
      required: [userId, name, points, streak, tasksCompleted, isCurrentUser]
    LogStudySessionRequest:
      type: object
      properties:
        durationSeconds:
          type: integer
        date:
          type: string
          format: date
      required: [durationSeconds]
    StudySessionResult:
      type: object
      properties:
        sessionMinutes:
          type: integer
        totalMinutesToday:
          type: integer
        goalMinutes:
          type: integer
        message:
          type: string
        goalReached:
          type: boolean
      required: [sessionMinutes, totalMinutesToday, goalMinutes, message, goalReached]
...