# lib/api-zod/src/generated/types/internshipListing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduEarn Student Productivity & Earning API
 * OpenAPI spec version: 0.2.0
 */
import type { InternshipListingType } from "./internshipListingType";

export interface InternshipListing {
  id: number;
  employerId: number;
  employerName: string;
  title: string;
  description: string;
  type: InternshipListingType;
  payment: number;
  paymentLabel: string;
  location: string;
  applyLink?: string | null;
  applyEmail?: string | null;
  skills?: string | null;
  createdAt: Date;
}